requirements.txt
streamlit
ultralytics
transformers
torch
opencv-python-headless
pillow
numpy
watchdog
